% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSmkFrames.R
\name{buildQuasiObsForApplication}
\alias{buildQuasiObsForApplication}
\title{Expand discrete time survival data into quasi observations on application data.}
\usage{
buildQuasiObsForApplication(d, observationWindowWidth, idColumnName,
  indexColumnName, parallelCluster = NULL)
}
\arguments{
\item{d}{data.frame to expand}

\item{observationWindowWidth}{scalar, step number to expand to}

\item{idColumnName}{character scalar, column to write original row ids in}

\item{indexColumnName}{character scalar, column to write quasi event indices into}

\item{parallelCluster, }{if not NULL parallel cluster to perform the work.}
}
\description{
This function assumes time starts at 1 and counts unit increments.  Expands into a regular
pattern of quasi observations up to step observationWindowWidth for scoring of data (without
known outcomes).
}
\examples{

d <- data.frame(lifetime=c(2,1,2),censored=c(FALSE,FALSE,TRUE))
buildQuasiObsForApplication(d,5,'origRow','sampleAge')

}

