% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSmkFrames.R
\name{buildQuasiObsForComparison}
\alias{buildQuasiObsForComparison}
\title{Expand discrete time survival data into quasi observations for holdout.}
\usage{
buildQuasiObsForComparison(d, observationWindowWidth, numberOfObservations,
  eventIndex, idColumnName, indexColumnName, eventColumnName,
  parallelCluster = NULL)
}
\arguments{
\item{d}{data.frame to expand}

\item{observationWindowWidth}{scalar, step number to expand to}

\item{numberOfObservations}{vector (entries >=1) for each row of d up to what step number did we observe}

\item{eventIndex}{vector for each row of d at what time did event occur.  Event i is considered censored if eventIndex[[i]] is NA or out of range of 1:numberOfObservations.}

\item{idColumnName}{character scalar, column to write original row ids in}

\item{indexColumnName}{character scalar, column to write quasi event indices into}

\item{eventColumnName}{character scalar, column to write quasi events indicator into.  If NA, write no such column.}

\item{parallelCluster, }{if not NULL parallel cluster to perform the work.}
}
\description{
This function assumes time starts at 1 and counts unit increments.  The returned
frame has the actual event data and expands to observationWindowWidth to imitiate
future application.  Quasi observation events past event or censoring are NA.
}
\examples{

d <- data.frame(lifetime=c(2,1,2),censored=c(FALSE,FALSE,TRUE))
buildQuasiObsForComparison(d,5,d$lifetime,ifelse(d$censored,NA,d$lifetime),
   'origRow','sampleAge','deathEvent')

}

