% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSmkFrames.R
\name{buildQuasiObsForTraining}
\alias{buildQuasiObsForTraining}
\title{Expand discrete time survival data into quasi observations for model training.}
\usage{
buildQuasiObsForTraining(d, numberOfObservations, eventIndex, idColumnName,
  indexColumnName, eventColumnName, parallelCluster = NULL, targetSize = NA)
}
\arguments{
\item{d}{data.frame to expand}

\item{numberOfObservations}{vector (entries >=1) for each row of d up to what step number did we observe}

\item{eventIndex}{vector for each row of d at what time did event occur.  Event i is considered censored if eventIndex[[i]] is NA or out of range of 1:numberOfObservations.}

\item{idColumnName}{character scalar, column to write original row ids in}

\item{indexColumnName}{character scalar, column to write quasi event indices into}

\item{eventColumnName}{character scalar, column to write quasi events indicator into.  If NA, write no such column.}

\item{parallelCluster, }{if not NULL parallel cluster to perform the work.}

\item{targetSize}{numeric scalar, if not NA target size for a uniform sample of the quasi observations.  This should be a very large number.}
}
\description{
This function assumes time starts at 1 and counts unit increments.
}
\examples{

d <- data.frame(lifetime=c(2,1,2),censored=c(FALSE,FALSE,TRUE))
buildQuasiObsForTraining(d,d$lifetime,ifelse(d$censored,NA,d$lifetime),
   'origRow','sampleAge','deathEvent')

}

