% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSurvival.R
\docType{package}
\name{QSurvival}
\alias{QSurvival}
\alias{QSurvival-package}
\title{QSurvival: simple survival models through quasi observations}
\description{
QSurvival converts survival modeling problem data into a format usable by standard machine learning methods.
QSurvival also includes methods for preparing the standard summaries used in survival analysis.
}
\details{
To learn more about QSurvival, start with the vignettes:
\code{browseVignettes(package = "QSurvival")}
}

