% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSorganizeResults.R
\name{summarizeActual}
\alias{summarizeActual}
\title{Calculate what fraction of ages are below each threshold. Do NOT use this on censored data (as all ages are interpreted as end).}
\usage{
summarizeActual(ages, range)
}
\arguments{
\item{ages}{numeric vector non-negative with integer values}

\item{range}{integer scalar integer posotive range to calculate to}
}
\value{
data frame with fraction less than our equal to each value
}
\description{
Calculate what fraction of ages are below each threshold. Do NOT use this on censored data (as all ages are interpreted as end).
}
\examples{

summarizeActual(1:5,7)

}

