% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSorganizeResults.R
\name{summarizeActualFrame}
\alias{summarizeActualFrame}
\title{Build observed survival curves for a frame.  Do NOT use this on censored data (as all ages are interpreted as end).}
\usage{
summarizeActualFrame(d, groupColumnName, ageColumnName,
  parallelCluster = NULL)
}
\arguments{
\item{d}{data.frame}

\item{groupColumnName}{name of grouping column}

\item{ageColumnName}{name of age column}

\item{parallelCluster}{optional parallel cluster to run on}
}
\value{
survival curves
}
\description{
Build observed survival curves for a frame.  Do NOT use this on censored data (as all ages are interpreted as end).
}
\examples{

s <- summarizeActualFrame(data.frame(age=1:10,
                                     group=as.factor((1:10)\%\%2)),
   'group','age')
# ggplot() + geom_line(data=s,mapping=aes(x=age,y=survival,color=group))

}

