% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSorganizeResults.R
\name{summarizeHazard}
\alias{summarizeHazard}
\title{Convert hazard to survival function and instantaneous
event intensities.}
\usage{
summarizeHazard(d, idColumnName, indexColumnName, hazardColumnName,
  survivalColumnName = "survival",
  deathIntensityColumnName = "deathIntensity", parallelCluster = NULL)
}
\arguments{
\item{d}{data.frame}

\item{idColumnName}{character scalar, column containing original row ids}

\item{indexColumnName}{character scalar, column containing quasi event indices}

\item{hazardColumnName}{character scalar, column containing hazard scores}

\item{survivalColumnName}{character scalar, column to write survival probability in}

\item{deathIntensityColumnName}{character scalar, column to write death intensities}

\item{parallelCluster}{optional parallel cluster to run on}
}
\value{
list, details=survival data.frame, expectedLifetime has lifetime estimates
}
\description{
Sums up hazard function grouped by original ID variable ordered by step.
}
\examples{

d <- data.frame(lifetime=c(2,1,2),censored=c(FALSE,FALSE,TRUE))
d2 <- buildQuasiObsForComparison(d,5,d$lifetime,ifelse(d$censored,NA,d$lifetime),
   'origRow','sampleAge','deathEvent')
d2$hazardPred <- 0.1
summarizeHazard(d2,'origRow','sampleAge','hazardPred',
   survivalColumnName='survival',
   deathIntensityColumnName='deathIntensity')


}

