% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSmkFrames.R
\name{buildQuasiObs}
\alias{buildQuasiObs}
\title{Expand discrete time survival data into quasi observations.}
\usage{
buildQuasiObs(d, observationWindowWidth, numberOfObservations, eventIndex,
  idColumnName, indexColumnName, eventColumnName, naAfterEvent, parallelCluster,
  targetSize)
}
\arguments{
\item{d}{data.frame to expand}

\item{observationWindowWidth}{scalar, minimum step number to expand to}

\item{numberOfObservations}{vector (entries >=1) for each row of d up to what step number did we observe}

\item{eventIndex}{vector for each row of d at what time did event occur.  Event i is considered censored if eventIndex[[i]] is NA or out of range of 1:numberOfObservations.}

\item{idColumnName}{character scalar, column to write original row ids in}

\item{indexColumnName}{character scalar, column to write quasi event indices into}

\item{eventColumnName}{character scalar, column to write quasi events indicator into.  If NA, write no such column.}

\item{naAfterEvent}{logical scalar, if TRUE write NA after event step in eventColumnName.}

\item{parallelCluster, }{if not NULL parallel cluster to perform the work.}

\item{targetSize}{numeric scalar, if not NA target size for a uniform sample of the quasi observations.}
}
\description{
This function assumes time starts at 1 and counts unit increments.
}

